Renamed system.css to webOS.css for a less misleading name
Renamed system.css to webOS.css for a less misleading name. Later in the project I will be adding a webOS API for programs without the OS to use to have better integration with the system and the user will import styles and Javascript, this will better help with that. I hope.